feat: add remove real time

+++ app/api/conversations/[conversationId]/route.ts
@@ -2,6 +2,7 @@ import getCurrentUser from "@/app/actions/getCurrentUser";
 import { NextResponse } from "next/server";
 
 import prisma from "@/app/libs/prismadb";
+import { pusherServer } from "@/app/libs/pusher";
 
 interface IParams {
   conversationId?: string;
@@ -19,6 +20,19 @@ export async function DELETE(
       return NextResponse.json(null);
     }
 
+    const existingConversation = await prisma.conversation.findUnique({
+      where: {
+        id: conversationId
+      },
+      include: {
+        users: true
+      }
+    });
+
+    if (!existingConversation) {
+      return new NextResponse('Invalid ID', { status: 400 });
+    }
+
     const deletedConversation = await prisma.conversation.deleteMany({
       where: {
         id: conversationId,
@@ -26,6 +40,12 @@ export async function DELETE(
           hasSome: [currentUser.id]
         },
       },
+    });
+
+    existingConversation.users.forEach((user) => {
+      if (user.email) {
+        pusherServer.trigger(user.email, 'conversation:remove', existingConversation);
+      }
     });
 
     return NextResponse.json(deletedConversation)
